// data/github/current/ui/packages/actions-secrets-variables-list/types.ts
export const ItemsScope = {
  Organization: 'Organization',
  OrgHostSetup: 'Host setup',
  OrgDevEnv: 'Development environment',
  OrgPrivateRegistry: 'Private registry',
  Repository: 'Repository',
  Environment: 'Environment',
  CodespaceUser: 'Codespace user',
} as const

export type ItemsScope = (typeof ItemsScope)[keyof typeof ItemsScope]

export const ListMode = {
  secret: 'secret',
  variable: 'variable',
} as const

export type ListMode = (typeof ListMode)[keyof typeof ListMode]

export interface RowData {
  // ID is the unique identifier for the row - same as name except for environment secrets on the repo page
  id: string
  name: string
  displayName: string
  scope?: ItemsScope
  environment?: string
  environmentUrl?: string
  editUrl?: string
  deleteUrl?: string
  lastUpdated?: Date
  visibility?: string
  override?: boolean
  value?: string
  registryUrl?: string
  registryType?: string
}

export interface Item {
  name: string
  environment_name?: string
  environment_url?: string
  visibility_description?: string
  updated_at: Date
  scope: ItemsScope
  value?: string
  registry_data?: {type: string; url: string}
}

export const FormMode = {
  add: 'add',
  update: 'update',
} as const

export type FormMode = (typeof FormMode)[keyof typeof FormMode]

export interface CrudUrls {
  editUrls?: {[secretName: string]: string}
  deleteUrls?: {[secretName: string]: string}
}

export interface TableActionProps {
  url?: string
  mode: ListMode
  message: string
  useDialog?: boolean
  isEditableInScope: boolean
  publicKey?: string
  keyId?: string
  tableDataUpdater: (item: Item, url: string, mode: FormMode) => void
}

export interface Description {
  text: string
  contextUrl?: string
  contextLinkText?: string
}

export interface ActionsSecretsVariablesListProps {
  items: Item[]
  overrides?: Item[]
  scope: ItemsScope
  environmentUrls?: {[environmentName: string]: string}
  crudUrls?: CrudUrls
  tableActionProps: TableActionProps
  mode: ListMode
  description?: Description
  isPrivateRepoInFreeOrg: boolean
  hideName?: boolean
  initialSortColumn?: string
}

export interface FormDialogProps {
  mode: ListMode
  name?: string
  value?: string
  onClose: () => void
  tableDataUpdater: (item: Item, url: string, mode: FormMode) => void
  formMode: FormMode
  url?: string
  publicKey?: string
  rowId: string
  keyId?: string
}

export interface DeleteDialogProps {
  url: string
  rowId: string
  mode: ListMode
  additionalFormData?: {[key: string]: string}
  setClose: (open: boolean, rowId: string) => void
  tableDataUpdater: (itemName: string) => void
}

export interface DescriptionElementProps {
  description: Description
}

export interface ActionsSecretsVariablesBlankstateProps {
  scope: ItemsScope
  mode: ListMode
  description: Description
  tableAction: JSX.Element
  isEditableInScope: boolean
  isPrivateRepoInFreeOrg: boolean
}